backend/requirements.txt
Flask
Flask-CORS
pymongo
python-dotenv
google-generativeai
werkzeug
gunicorn
